+++ libraries/ACS712/test/unit_test_001.cpp
@@ -212,4 +212,4 @@ unittest(test_AmperePerStep)
 unittest_main()
 
 
-// -- END OF FILE --
+//  -- END OF FILE --